+++ src/app/core/database/multi-tab-detection.service.ts
@@ -184,33 +184,3 @@ export class MultiTabDetectionService implements OnDestroy {
     this.isMultipleTabsOpen$.complete();
   }
 }
-
-/**
- * Detect IndexedDB/PouchDB corruption symptoms commonly observed when the app
- * is used in multiple tabs with concurrent writes.
- *
- * We intentionally match on `transaction was aborted` because this is the
- * reliable signal present on the actual error object propagated to app code
- * after IndexedDB global-failure cases. The `unknown_error` label is too
- * generic and may appear for unrelated failures.
- */
-export function isKnownMultiTabDatabaseCorruption(error: unknown): boolean {
-  const text = errorToText(error).toLowerCase();
-  const transactionAbortedMatch = text.includes("transaction was aborted");
-  const globalFailureMatch = text.includes("database has a global failure");
-  const constraintSeqMatch =
-    text.includes("constrainterror") && text.includes("seq");
-  const isKnownCorruption =
-    transactionAbortedMatch || globalFailureMatch || constraintSeqMatch;
-
-  return isKnownCorruption;
-}
-
-function errorToText(error: unknown): string {
-  if (error instanceof Error) {
-    const errorWithExtras = error as Error & Record<string, unknown>;
-    return `${error.name} ${error.message} ${JSON.stringify(errorWithExtras)}`;
-  }
-  if (typeof error === "string") return error;
-  return JSON.stringify(error ?? "");
-}
