+++ src/app/core/database/pouchdb/pouchdb-corruption-recovery.service.ts
@@ -6,6 +6,7 @@ import {
 } from "#src/app/core/common-components/confirmation-dialog/confirmation-dialog/confirmation-dialog.component";
 import { LOCATION_TOKEN } from "#src/app/utils/di-tokens";
 import { BackupService } from "../../admin/backup/backup.service";
+import { Logging } from "../../logging/logging.service";
 
 /**
  * Handles user-facing recovery flows when PouchDB IndexedDB corruption is detected.
@@ -72,4 +73,44 @@ We are working on improvements to allow this in the future.`,
     sessionStorage.setItem(BackupService.RESET_PENDING_KEY, "1");
     this.location.pathname = "";
   }
+
+  handleKnownMultiTabCorruption(err: unknown, logMessage: string): void {
+    if (!isKnownMultiTabDatabaseCorruption(err)) {
+      return;
+    }
+    Logging.warn(logMessage, err);
+    this.promptResetApplicationDialog().catch((callbackError) =>
+      Logging.warn("onKnownMultiTabCorruption callback failed", callbackError),
+    );
+  }
+}
+
+/**
+ * Detect IndexedDB/PouchDB corruption symptoms commonly observed when the app
+ * is used in multiple tabs with concurrent writes.
+ *
+ * We intentionally match on `transaction was aborted` because this is the
+ * reliable signal present on the actual error object propagated to app code
+ * after IndexedDB global-failure cases. The `unknown_error` label is too
+ * generic and may appear for unrelated failures.
+ */
+export function isKnownMultiTabDatabaseCorruption(error: unknown): boolean {
+  const text = errorToText(error).toLowerCase();
+  const transactionAbortedMatch = text.includes("transaction was aborted");
+  const globalFailureMatch = text.includes("database has a global failure");
+  const constraintSeqMatch =
+    text.includes("constrainterror") && text.includes("seq");
+  const isKnownCorruption =
+    transactionAbortedMatch || globalFailureMatch || constraintSeqMatch;
+
+  return isKnownCorruption;
+}
+
+function errorToText(error: unknown): string {
+  if (error instanceof Error) {
+    const errorWithExtras = error as Error & Record<string, unknown>;
+    return `${error.name} ${error.message} ${JSON.stringify(errorWithExtras)}`;
+  }
+  if (typeof error === "string") return error;
+  return JSON.stringify(error ?? "");
 }
