+++ src/app/core/database/pouchdb/synced-pouch-database.ts
@@ -19,8 +19,9 @@ import { EMPTY, from, interval, merge, of } from "rxjs";
 import { LoginState } from "../../session/session-states/login-state.enum";
 import { NotAvailableOfflineError } from "../../session/not-available-offline.error";
 import { AlertService } from "../../alerts/alert.service";
-import { isKnownMultiTabDatabaseCorruption } from "../multi-tab-detection.service";
+import { isKnownMultiTabDatabaseCorruption } from "./pouchdb-corruption-recovery.service";
 import { QueryOptions } from "../database";
+import { PouchdbCorruptionRecoveryService } from "./pouchdb-corruption-recovery.service";
 
 /**
  * An alternative implementation of PouchDatabase that additionally
@@ -42,7 +43,7 @@ export class SyncedPouchDatabase extends PouchDatabase {
   private readonly navigator: Navigator;
   private readonly loginStateSubject: LoginStateSubject;
   private readonly alertService?: AlertService;
-  private readonly onKnownMultiTabCorruption?: () => void | Promise<void>;
+  private readonly corruptionRecovery?: PouchdbCorruptionRecoveryService;
   private remoteDatabase: RemotePouchDatabase;
   private syncState: SyncStateSubject = new SyncStateSubject();
 
@@ -70,13 +71,13 @@ export class SyncedPouchDatabase extends PouchDatabase {
     loginStateSubject: LoginStateSubject,
     ngZone?: NgZone,
     alertService?: AlertService,
-    onKnownMultiTabCorruption?: () => void | Promise<void>,
+    corruptionRecovery?: PouchdbCorruptionRecoveryService,
   ) {
     super(dbName, globalSyncState, ngZone);
     this.navigator = navigator;
     this.loginStateSubject = loginStateSubject;
     this.alertService = alertService;
-    this.onKnownMultiTabCorruption = onKnownMultiTabCorruption;
+    this.corruptionRecovery = corruptionRecovery;
 
     this.remoteDatabase = new RemotePouchDatabase(
       dbName,
@@ -227,7 +228,7 @@ export class SyncedPouchDatabase extends PouchDatabase {
             err,
           );
         } else if (isKnownMultiTabDatabaseCorruption(err)) {
-          this.handleKnownMultiTabCorruption(
+          this.corruptionRecovery?.handleKnownMultiTabCorruption(
             err,
             `sync failed [${this.dbName}]: likely multi-tab IndexedDB corruption. Last synced batch: [${lastSyncedDocIds.join(", ")}]`,
           );
@@ -248,7 +249,7 @@ export class SyncedPouchDatabase extends PouchDatabase {
     try {
       return await super.put(object, forceOverwrite);
     } catch (err) {
-      this.handleKnownMultiTabCorruption(
+      this.corruptionRecovery?.handleKnownMultiTabCorruption(
         err,
         `put failed [${this.dbName}]: likely multi-tab IndexedDB corruption`,
       );
@@ -261,7 +262,7 @@ export class SyncedPouchDatabase extends PouchDatabase {
     options: QueryOptions,
   ): Promise<any> {
     return super.query(fun, options).catch((err) => {
-      this.handleKnownMultiTabCorruption(
+      this.corruptionRecovery?.handleKnownMultiTabCorruption(
         err,
         `query failed [${this.dbName}]: likely multi-tab IndexedDB corruption`,
       );
@@ -276,25 +277,6 @@ export class SyncedPouchDatabase extends PouchDatabase {
       message.includes("IDBObjectStore") ||
       message.includes("Failed to execute")
     );
-  }
-
-  private handleKnownMultiTabCorruption(
-    err: unknown,
-    logMessage: string,
-  ): void {
-    if (!isKnownMultiTabDatabaseCorruption(err)) {
-      return;
-    }
-    Logging.warn(logMessage, err);
-    const callbackResult = this.onKnownMultiTabCorruption?.();
-    if (callbackResult) {
-      callbackResult.catch((callbackError) =>
-        Logging.warn(
-          "onKnownMultiTabCorruption callback failed",
-          callbackError,
-        ),
-      );
-    }
   }
 
   /**
